Disconnect build.moxie parent from parentPom
In order to create a jar than can be synced to Maven Central using Sonatype OSS
Sync, the artifact's pom must have org.sonatype.oss:oss-parent:N as its parent.

This change creates a new key (parentPom) which allows you to specify a parentPom
and eliminates the special case code for controlling generation of the parent section
of the pom.