Replaced dependency repository boost with repository affinities
The dependency  boost parameter, while seemingly a good idea, didn't work well in practice.  It only worked for the specified dependency and not for the dependency's graph.  For example, defining a jgit dependency also pulls in a jgit-parent dependency.  The parent, however, would not inherit the boost - not should it.

In place of the boost parameter, a repository definition may now specify groupId:artifactId or groupId or startsWith values to boost that repository for a dependency.