Fixed test->build classpath dependency leaking
This bug would manifest if the project being built specified a newer,
conflicting version of one of Moxie's dependencies.  This would result
in a classpath failure when running the Moxie task because the newer
project dependency was already on the classpath.

The solution was to stop leaking the test classpath into the build
classpath with two important exceptions: JUnit and TestNG which both
need to be loadable on the build classpath for MxTest.