Update Ant to 1.10.13 - requires Java 8
Update Ant to the latest version: 1.10.13
This version requires Java 8 as the minimum Java version.
I believe this is a valid minimum version and support for versions below
Java 8 is no longer required so we go with the 1.10 line instead of sticking
to the 1.9 line.

Thus, the javac task is also switched to Java 8.

Due to the change of the minimum Java version, the minor version increases.