Generating Pom index reads POMs with loose requirements
The `readAllPoms` method in the `IMavenCache` is only used in the
method to generate the POM index. But for that we don't need the POM
reader to resolve properties and read parent POMs. In the repository
we may not have all parent POMs, but we still want the deployed artifacts
in the index. So use the LOOSE requirements when reading POMs.