Add `keep` parameter to `ghpages` task
Add a `keep` parameter as a nested element to the `ghpages` task, which
allows to specify files that should be kept from the previous tree, when
the `obliterate` parameter is set to True. While we may want to obliterate
all pages that are generated, there are some, like Github's CNAME, that
are not generated and need to be kept around between commits.